requirements.txt
anthropic>=0.40.0
python-dotenv==1.0.0
PyPDF2==3.0.1
reportlab==4.0.9
requests==2.31.0
pyperclip==1.8.2
beautifulsoup4==4.12.3
flask==3.0.0
flask-cors==4.0.0
